audio: host: add host folder for host source file
Move all host related code to this folder.
